Remove -Wno-sign-compare
Former-commit-id: 4e9642dd673067cc48bf970995483a15589e49c3

+++ src/bytecode/ByteCode.cpp
@@ -280,7 +280,7 @@ void dumpBytecode(CodeBlock* codeBlock)
             printf("t2: %d,", ex->m_targetIndex1);
         }
 
-        for (int i = 0; i < ex->m_sourceIndexes.size() ; i ++) {
+        for (size_t i = 0; i < ex->m_sourceIndexes.size() ; i ++) {
             printf("s: %d,", ex->m_sourceIndexes[i]);
         }
         printf("]");
